Move auth into recommendations controller temp
